feat: 🎸 audio and visual patch updates (messy)
Overdue commit of many changes in the Playground files where I've been
testing audio and visual synthesis. Includes addition of lodash utility
(used in the Euclidean rhythms algorithm). Plan is for this commit to be
an archival entry. From there I intend to clean things up and commit
more often.